# app/src/main/openapi-generated/com/ecommerce/openapi/apis/OrderApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.ecommerce.openapi.apis;

import java.math.BigDecimal;
import com.ecommerce.openapi.apis.models.CartProduct;
import com.ecommerce.openapi.apis.models.CartRegist;
import com.ecommerce.openapi.apis.models.OrderProduct;

import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;


import javax.validation.Valid;
import java.util.List;

@Validated

@RestController
public interface OrderApi {

    /**
     * POST /cart/order/{id} : 장바구니에 담긴 상품을 주문.
     *
     * @param id 상품주문할 user의 id (required)
     * @return 주문완료된 상품 리스트와 결과 (status code 200)
     *         or unexpected error (status code 400)
     */

    @PostMapping(
        value = "/cart/order/{id}",
        produces = { "application/json" }
    )
    ResponseEntity<OrderProduct> orderCartProduct(@PathVariable("id") BigDecimal id);


    /**
     * POST /cart/regist : 주문하기 위한 상품을 장바구니에 담기.
     *
     * @param cartRegist 추가할 cart의 정보 (optional)
     * @return cart에 담긴 상품 리스트 (status code 200)
     *         or unexpected error (status code 400)
     */

    @PostMapping(
        value = "/cart/regist",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<List<CartProduct>> registCartProduct( @Valid @RequestBody(required = false) CartRegist cartRegist);


    /**
     * GET /cart/{id} : 장바구니에 담긴 user의 상품리스트.
     *
     * @param id cart정보를 가져올 user의 id (required)
     * @return user의 cart 리스트 (status code 200)
     *         or unexpected error (status code 400)
     */

    @GetMapping(
        value = "/cart/{id}",
        produces = { "application/json" }
    )
    ResponseEntity<List<CartProduct>> searchCartByUser(@PathVariable("id") BigDecimal id);

}