# app/src/main/openapi-generated/com/ecommerce/openapi/apis/ProductApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.ecommerce.openapi.apis;

import java.math.BigDecimal;

import com.ecommerce.openapi.apis.models.FamousProduct;
import com.ecommerce.openapi.apis.models.Product;

import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;


import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;

@Validated

@RestController
public interface ProductApi {

    /**
     * GET /products/famous : 최근 3일간 가장 많이 팔린 상위 5개 상품을 조회함.
     *
     * @param date 최근 3일의 기준 날짜 (required)
     * @return 조건에 맞는 상품 리스트 (status code 200)
     *         or unexpected error (status code 400)
     */

    @GetMapping(
        value = "/products/famous",
        produces = { "application/json" }
    )
    ResponseEntity<List<FamousProduct>> searchFamousProducts(@NotNull  @Valid @RequestParam(value = "date", required = true) String date);


    /**
     * GET /products : 상품을 조회함.
     *
     * @param page 페이지네이션의 인덱스 (required)
     * @param rowsPerPage 페이지별 row (required)
     * @return 조건에 맞는 상품 리스트 (status code 200)
     *         or unexpected error (status code 400)
     */

    @GetMapping(
        value = "/products",
        produces = { "application/json" }
    )
    ResponseEntity<List<Product>> searchProducts(@NotNull  @Valid @RequestParam(value = "page", required = true) BigDecimal page,@NotNull  @Valid @RequestParam(value = "rowsPerPage", required = true) BigDecimal rowsPerPage);

}